Rename wave-field to field-based and update imports
Renamed files and updated terminology from 'wave-field' to 'field-based' throughout the codebase and documentation for consistency. Adjusted all relevant imports, docstrings, and UI labels to reflect the new naming convention. This improves clarity between the Level-0 granule-based and Level-1 field-based computational approaches.
