+++ openwave/__init__.py
@@ -1,7 +1,7 @@
 """Core package for the OpenWave project.
 
-OpenWave provides tools to simulate the formation and behavior of matter and
-energy based on Energy Wave Theory.
+OpenWave is an open-source computational physics toolkit for modeling matter and energy phenomena using wave-field dynamics.
+
 """
 
 __version__ = "0.3.1"
