+++ openwave/validations/stability_analysis_springmass.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from openwave.common import constants, equations
 
-import openwave.spacetime.medium_level0 as medium
+import openwave.spacetime.L0_granule_based as data_grid
 
 
 # From constants
@@ -18,7 +18,7 @@ STIFFNESS = 1e-13  # N/m (already reduced!)
 granules_per_edge = TARGET_GRANULES ** (1 / 3) * 0.8  # 80% fill factor
 unit_cell_edge = UNIVERSE_EDGE / granules_per_edge
 # unit_cell_edge = 2 * np.e * constants.PLANCK_LENGTH  # ~5.4e-35 m, override to Planck scale
-granule = medium.BCCGranule(unit_cell_edge, UNIVERSE_EDGE)
+granule = data_grid.BCCGranule(unit_cell_edge, UNIVERSE_EDGE)
 
 # Spring-mass system natural frequency
 frequency = equations.compute_natural_frequency(STIFFNESS, granule.mass)  # Hz
